==> R/00_all.qmd <==
---
title: "00_all.qmd"
format: html
editor: visual
---

#Run all scripts

```{r}
#Run all scripts

#Load
rmarkdown::render(input = '01_load.qmd',

                  output_dir = '../results')
#clean
rmarkdown::render(input = '02_clean.qmd',

                  output_dir = '../results')
#Augment
rmarkdown::render(input = '03_augment.qmd',

                  output_dir = '../results')
#Describe
rmarkdown::render(input = '04_describe.qmd',

                  output_dir = '../results')
#Analysis_1
rmarkdown::render(input = '05_analysis_1.qmd',

                  output_dir = '../results')
#Analysis_2
rmarkdown::render(input = '06_analysis_2.qmd',

                  output_dir = '../results')

```
