==> R/01_load.qmd <==
---
title: "01_load"
format: html
editor: visual
---

# Load

## Load libraries

```{r message=FALSE}
library(tidyverse)
library(readxl)
library(lubridate)
library(dplyr)
library(janitor)
library(data.table)
library(purrr)
library(cowplot)
library(broom) 
library(ggplot2)
library(table1)
library(reshape2)
library(RColorBrewer)
library(survival)
library(survminer)
library(gridExtra)
library(viridis)
library(knitr)
```

## Load raw_data.xlsx from \_raw folder

```{r}
# Load the data using readxl library and remove missing values present in different formats

file_path = "~/group_18_project/_raw/raw_data.xlsx"

# Load data from sheet 1
dat_load1 <- read_xlsx(path = file_path, na = c("unknown", "Unknown", "N/A", "Unkown", "unkown", "None", "none", "" ), sheet = 1)

# Load data from sheet 2
dat_load2 <- read_excel(path = file_path, sheet = 2)

# Load data from sheet 3
dat_load3 <- read_excel(path = file_path, sheet = 3)
```

## Save the data as 01_dat_load.tsv in data folder

```{r}
# Save data from sheet 1 
write.table(x = dat_load1, "~/group_18_project/data/01_dat_load1.tsv", 
            sep = "\t", 
            quote = TRUE,
            row.names = FALSE)

# Save data from sheet 2
write.table(x = dat_load2, 
            file = "~/group_18_project/data/01_dat_load2.tsv", 
            sep = "\t", 
            quote = TRUE,
            row.names = FALSE)

# Save data from sheet 3 
write.table(x = dat_load3, 
            file = "~/group_18_project/data/01_dat_load3.tsv", 
            sep = "\t", 
            quote = TRUE,
            row.names = FALSE)
```
