==> R/04_describe.qmd <==
---
title: "03_describe.qmd"
format: html
editor: visual
---

# Describe

The dataset under consideration originates from an in-depth exploration of medical records pertaining to individuals who underwent radiation therapy for prostate cancer within the timeframe of January 2000 to September 2015, specifically within two Detroit Medical Center hospitals. The dataset encompasses information on 709 patients. The primary objective of this investigation was to assess the frequency and severity of radiation-induced hemorrhagic cystitis (HC) in patients subjected to radiation therapy.

The findings reveal that approximately 11.1% of the patients in the dataset developed HC subsequent to their radiation therapy. The analysis delved into various factors, including the type of radiation administered, the overall health status of the patients, and the specifics of their treatment. Interestingly, no significant discrepancy was observed based on the type of radiation. On average, the onset of HC occurred around 79 months post-radiation, accompanied by considerable medical complexities such as multiple hospital visits, blood transfusions, and, in certain instances, intricate medical procedures.

Demographically, the majority of these patients were of African American descent, with an average age of around 69 at the time of radiation. The spectrum of complications ranged from urinary tract infections to more severe conditions like kidney issues, bladder ruptures, and unfortunately, instances of mortality.

The dataset discussed in the article comprises 3 files. The initial file stands out as the primary source of information, containing the most extensive dataset. Following a thorough cleaning process, this file underwent augmentation. In contrast, the second and third files did not necessitate augmentation, having been deemed complete after cleaning and merged together. Therefore, it is important to provide descriptions of the `dat_clean2` file as well, rather than focusing solely on the augmented one.

## dat_clean2

### Load dat_clean2 from data folder

```{r}
dat_clean2 <- read_tsv(
  file = "~/group_18_project/data/02_dat_clean2.tsv",
  show_col_types = FALSE)
```

### Inspect the dat_clean2

```{r}

dat_clean2 |> head(10)
```

The data from the dataset `dat_clean2` shows the information of all 739 patients who underwent radiation therapy for prostate cancer, with both the patients who developed HC and the patients who didn't. This dataset has fewer columns, only giving the `hospital_id`, `hc_incidence`, `bladder_cancer`, `prostate_cancer` and the `treatment`. The 2 most important variables in this dataset are the `hc_incidence`, stating if the patient developed HC or not, and the `treatment`, showing the type of radiation or other treatment they went through.

```{r}
dat_clean2 |>
  table1(x = formula(~ treatment + bladder_cancer | prostate_cancer),
         data = _)
```

The table shows the key statistics of the variables `treatment` and `bladder_cancer`. The most common treatments are EBRT at 34.1% and Brachytherapy at 6.0%. Other treatments include Adjuvant therapy (3.7%), IMRT (14.7%), Unknown treatments (25.2%), and a combination of Neutron/Photon therapies with EBRT or IMRT. The majority of individuals (95.0%) in the dataset do not have bladder cancer, while a smaller proportion (5.0%) has bladder cancer.

#### Key plot 1

The presented plot illustrates the distribution of patients who experienced radiation-induced Hemorrhagic Cystitis (HC) based on the type of treatment received. Utilizing proportions in this visualization offers a comparative view of how different treatments are distributed relative to the total number of cases with HC incidence in the `dat_clean2` dataset. Each bar represents the proportion of patients receiving a specific treatment within the categories of HC incidence, providing valuable insights into the treatment patterns in the context of radiation-induced HC.

```{r}
plot1 <- 
  dat_clean2 |> 
  count(treatment, hc_incidence) |>
  group_by(treatment) |>
  mutate(freq = n / sum(n)) |>
  ggplot(aes(x = treatment, 
           y = freq,
           fill = hc_incidence)) +  
  geom_col(position = position_dodge(
    preserve = "single"),
    colour = "black",
    alpha = 0.5) +
  geom_hline(yintercept = 0) +
  labs(x = "Treatment Received",    
       y = "Proportion",              
       fill = "Study of Radiation-Induced Hemorrhagic Cystitis"     
  ) +
  theme_minimal(base_size = 12) +
  theme(legend.position = "top",
        panel.grid.major.x = element_blank(),
        axis.text.x = element_text(angle = 45,
                                   hjust = 1,
                                   vjust = 1.1))

ggsave(filename = "key_plot_1.png", 
      plot = plot1, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)
```

As seen in the plot, the majority of patients did not develop HC. While a small percentage of patients did develop HC in all the radiation types, brachytherapy was the only radiation therapy without HC development.

## dat_aug

### Load dat_aug from data folder

```{r}
dat_aug1 <- read_tsv(
  file = "~/group_18_project/data/03_dat_aug1.tsv",
  show_col_types = FALSE)
```

### Inspect the dat_aug1.tsv

```{r}
dat_aug1 |> head(10)
```

The dataset `dat_aug1` shows the information on only the 80 patients who developed HC subsequent to their radiation therapy. The primary focus of the analysis is to assess the frequency and severity of radiation-induced hemorrhagic cystitis (HC) in these patients. Standard information of each patient is given such as: `date_of_birth`, `age_at_radiation` and `ethnicity`, but the following variables were important to look at:

-   **`BMI`** derived from the patient's height and weight. BMI could be crucial for understanding the association between body composition and the development of HC, as research has suggested that BMI may influence treatment outcomes and complications in cancer patients.

-   **`cancer_type`** indicating the specific type or category of prostate cancer that patients are diagnosed with. Understanding the cancer type is essential for tailoring effective treatments and analyzing how different types may be associated with the occurrence and severity of radiation-induced complications (HC).

-   **`hematuria_grade`** indicating the severity or grade of hematuria experienced by patients. Hematuria grade is a qualitative assessment of the severity of blood in the urine. It ranges from mild to severe, reflecting the amount and visibility of blood. Understanding hematuria grade provides insights into the severity of complications, especially those induced by radiation therapy.

-   **`num_comorbidities`** representing the number of comorbid conditions a patient has. Comorbidities are additional medical conditions that coexist with the primary condition (prostate cancer). Assessing comorbidities is essential for understanding the overall health status of patients, as it can influence treatment decisions and outcomes.

-   **`radiation_type`** representing the types of radiation therapy administered to the patients. These include:

    -   EBRT (External Beam Radiation Therapy): high-energy X-rays are directed at the cancerous tissue from outside the body.

    -   IMRT (Intensity-Modulated Radiation Therapy): IMRT is an advanced form of EBRT that adjusts the intensity of the radiation beams.

    -   Neutron Photon: This combines the use of neutrons and photons (X-rays) for radiation treatment. Neutrons have unique properties that may enhance their effectiveness in certain cases and will be used for specific types of cancer.

-   **`num_days_admited`** indicates the number of days a patient was admitted to the hospital. Tracking hospital admission days provides insights into the impact of radiation therapy on patients' well-being, including potential complications and recovery periods.

-   **`num_hosp_complications`** denotes the number of hospital complications experienced by patients. Monitoring hospital complications is vital for assessing the overall safety and efficacy of radiation therapy. It helps identify potential risks associated with the treatment and informs healthcare decisions for patient care.

```{r}
dat_aug1 |>
  mutate(hematuria_grade = factor(hematuria_grade)) |>
  table1(x = formula(~ age_at_radiation + ethnicity + radiation_type | hematuria_grade),
         data = _)
```

The table shows the key statistics of the variables `age_at_radiation`, `ethnicity` and `radiation_type`. The dataset includes patients with a mean age at radiation of 68.7 years, a median age of 68.5 years, which ranges from 48 to 89 years. The ethnicity information is available for 90% of patients, with the majority being African American (92.5%) while Caucasians make up 7.5% of the dataset. The most common type of radiation is EBRT at 55% and different combinations of radiation types are also present, with varying frequencies.

### 

#### Key plot 2

The boxplot is used to visualize the distribution of the normalized values for various variables obtained for patient that developed HC. Each box represents the interquartile range (IQR), with the median depicted as a horizontal line inside the box. The plot provides a concise summary of the central tendency, spread, and potential outliers in the normalized data across different variables.

```{r warning = FALSE}

box_plot_data <- 
  dat_aug1 |>
  select(age_at_radiation,
         BMI, 
         num_comorbidities, 
         num_days_admitted, 
         num_hosp_complications, 
         number_of_admissions_with_urology_consultation, 
         number_of_cystoscopies, 
         number_transfusions, radiation_dosage,
         radiation_duration, 
         radiation_completion_to_first_hc) |>
  mutate_all(scale)

# Reshape data into long format
long_data <- 
  gather(box_plot_data, 
                    key = "Variable",
                    value = "Value")

# Create boxplots with normalized data, customized variable names
plot2 <- 
  ggplot(data =long_data, 
         aes(x = Variable, 
             y = Value, 
             fill = Variable)) +
  geom_boxplot(color = "black",
               alpha = 0.5) +
  labs(title = "Box plot visualization of selected variables",
       x = "Variables", 
       y = "Normalized Values") +
  theme_minimal() +
  theme(axis.text.x = element_blank()) 

print(plot2)

# Save the plot to folder results
ggsave(filename = "key_plot_2.png", 
      plot = plot2, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)

```

#### Key plot 3

The 4 box plots each show the number of patients with HC that had the following variables: Age distribution at radiation treatment for HC patients, BMI distribution of HC patients, Distribution of the number of comorbidities in HC patients and finally the distribution of the number of hospital complications in HC patients.

```{r warning = FALSE}

df = dat_aug1

# Age Distribution
plot_age <- 
  ggplot(df, aes(x = age_at_radiation)) +
  geom_histogram(binwidth = 5, fill = viridis(10)[1], color = "white", alpha = 0.7) +
  labs(#title = "Age Distribution of Patients with HC",
    x = "Age at Radiation") +
  theme_classic()

# BMI Distribution
plot_bmi <- 
  ggplot(df, aes(x = BMI)) +
  geom_histogram(binwidth = 2, fill = viridis(10)[2], color = "white", alpha = 0.7) +
  labs(#title = "BMI Distribution of Patients with HC",
    x = "BMI") +
  theme_classic()

# Number of Comorbidities Distribution
plot_comorbidities <- 
  ggplot(df, aes(x = num_comorbidities)) +
  geom_histogram(binwidth = 1, fill = viridis(10)[3], color = "white", alpha = 0.7) +
  labs(#title = "Distribution of Number of Comorbidities among Patients with HC",
    x = "Number of Comorbidities") +
  theme_classic()

# Number of Comorbidities Distribution
plot_hosp_comp <- 
  ggplot(df, aes(x = num_hosp_complications)) +
  geom_histogram(binwidth = 1, fill = viridis(10)[4], color = "white", alpha = 0.7) +
  labs(#title = "Distribution of Number of Comorbidities among Patients with HC",
    x = "Number of Complications") + 
  theme_classic()

# Arrange plots in a grid
grid.arrange(plot_age, plot_bmi, plot_comorbidities, plot_hosp_comp,  ncol = 2)

# Save the plot to folder results
ggsave(filename = "key_plot_3.png", 
      plot = plot_bmi, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)
```
