==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1.qmd"
format: html
editor: visual
---

# Analysis 1

## Load clean data from data folder

```{r}

analysis_1.1 <- read_tsv(
  file = "~/group_18_project/data/03_dat_aug1.tsv",
  show_col_types = FALSE)

analysis_1.2  <- read_tsv(
  file = "~/group_18_project/data/02_dat_clean2.tsv",
  show_col_types = FALSE)
```

## Key plot 4

Linear Regression Analysis of Body Mass Index (BMI) in relation to radiation dose.

From the plot below, we can observe that the higher the BMI, the higher the radiation dose to the patients. However, the quality of the data is not sufficient to confirm these as we had many NAs in the radiation dosage column.

```{r warning = FALSE}

# Create the plot
plot4 <- 
  analysis_1.1 |> 
  na.omit(analysis_1.1) |> 
  ggplot(aes(x = radiation_dosage,
             y = BMI)) +
  geom_point() +
  geom_smooth(method = "lm", 
              se = FALSE, 
              color = "blue") +
  theme_minimal(base_size = 12) +  # Adjusted base_size to match the previous code
  theme(legend.position = "top",
        panel.grid.major.x = element_blank(),
        axis.text.x = element_text(angle = 0,
                                   hjust = 1,
                                   vjust = 1.1)) +  # Applied the previous theme adjustments
  labs(x = "Radiation Dose",    
       y = "BMI",              
       title = "Linear Regression between BMI and radiation dose") +
  theme(plot.title = element_text(hjust = 0.5, size = 16))  

# Perform linear regression
model <- lm(radiation_dosage ~ BMI, data = analysis_1.1)

# Print regression summary
summary(model)


#Print plot
print(plot4)

# Save plot to folder called "results"
ggsave(filename = "key_plot_4.png", 
      plot = plot4, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)

```

## Key plot 5

This plot shows the Distribution of Radiation Doses across different radiation types. According to literature \[1\], a standardized threshold has been added distinguishing Low, Medium, and High Levels across various radiation types.

From the plot below, we can observe that 15% of patients underwent EBRT medium dose therapy, while 10% received IMRT medium dose therapy, making them the two most prevalent radiation treatments.

\[1\] Ganswindt, U., Paulsen, F., Anastasiadis, A. G., Stenzl, A., Bamberg, M., & Belka, C. (2005). 70 Gy or more: which dose for which prostate cancer?. Journal of cancer research and clinical oncology, 131(7), 407--419.

```{r message=FALSE}
 
# Define dose categories based on paper
low_dose_threshold <- 70
high_dose_threshold <- 85

# Categorize doses using mutate
analysis_1.1 <- 
  mutate(analysis_1.1, 
         dose_category = cut(radiation_dosage, 
                             breaks = c(-Inf,low_dose_threshold, high_dose_threshold, Inf),
                             labels = c('Low Dose', 'Medium Dose', 'High Dose')))

# Calculate percentages for each type of radiation using summarise
result_3 <- 
  analysis_1.1 |> 
  group_by(radiation_type, dose_category) |> 
  summarise(Proportion = n()/nrow(analysis_1.1)) |> 
  filter(!is.na(dose_category) & !is.na(radiation_type) & !is.na(Proportion)) |>  
  ungroup()

# Create a grouped plot with viridis color scheme
plot5 <- result_3 |> 
  ggplot(aes(x = radiation_type, 
             y = Proportion, 
             fill = dose_category)) +
  geom_bar(stat = "identity",
    position = position_dodge(preserve = "single"),
    color = "black",
    alpha = 0.5) + 
  geom_hline(yintercept = 0) +
  labs(title = "Radiation Type and corresponding dosage category",
       #x = "Type of Radiation",
       y = "Proportion",
       fill = "Radiation Dosage Category") +
  theme_minimal(base_size = 12) +
  theme(legend.position = "right",
        panel.grid.major.x = element_blank(),
        axis.text.x = element_text(angle = 45,
                                   hjust = 1,
                                   vjust = 1.08),
        axis.title.x = element_blank()
        )

# Print the plot
print(plot5)


# Save plot to the "results" folder
ggsave(filename = "key_plot_5.png", 
      plot = plot5, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)
```

## Key plot 6

Description: This boxplot compares the average time passed since the completion of radiation treatment and the first incident of HC. For this the mean of the time passed since the completion of radiation treatment and first incident of HC is calculated and presented on the y-axis while on the x-axis the different types of radiation is shown. This could give an insight of the impact of the radiation type on the occurrence of HC in the patients.

```{r}
# Select relevant columns and filter out rows with missing values
plot6_data <- 
  analysis_1.1 |>
  select(radiation_type, radiation_completion_to_first_hc) |>
  mutate(TimePassed = as.numeric(radiation_completion_to_first_hc)) |>
  filter(!is.na(TimePassed)) |> 
  filter(!is.na(radiation_type)) 

# Group the data by radiation type and calculate mean time passed
result_6 <- 
  plot6_data |>
  group_by(radiation_type) |>
  summarize(
    AvgTimePassed = mean(TimePassed),
    MedianTimePassed = median(TimePassed),
    TotalObservations = n())

# Create a bar plot comparing average time since radiation for different type
plot6 <- 
  ggplot(data = result_6, 
         aes(x = radiation_type,
             y = AvgTimePassed, 
             fill = radiation_type)) +
  geom_bar(stat = "identity", 
           position = "dodge", 
           color = "black", 
           alpha = 0.5) +
  labs(title = "Average Time of Radiation Completion to First HC Incident",
       x = "Type of Radiation",
       y = "Average Time Of Radiation To First HC\n(months)",
       fill = "Type of Radiation") +
  theme_minimal(base_size = 12) + 
  theme(axis.text.x = element_blank())

# Print the plot
print(plot6)

# Save plot to the "results" folder
ggsave(filename = "key_plot_6.png", 
      plot = plot6, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)

```

## Key plot 7

Relationship Between Body Mass Index (BMI) and time from radiation completion to first incidence of Hemorrhagic Cystitis.

From the plot below, we can observe that all Caucasian people developed Hemorrhagic cystitis before 100 months from the end of their cancer therapy. The vast majority of the patients got HC before 200 months and after them the cases of HC were rare.

```{r warning = FALSE}

#Create a scatter plot
plot7 <- 
  ggplot(data = analysis_1.1,
         aes(x = radiation_completion_to_first_hc, 
             y = BMI , 
             color = ethnicity)) +
  geom_point() +
  labs(title = "BMI vs Radiation Completion to First HC",
       x = "Radiation Completion to First HC",
       y = "BMI")+
  scale_color_manual(values = c("african_american" = "blue",
                                "caucasian" = "red"))+
  theme_minimal(base_size = 12)
  

#Print plot
print(plot7)

# Save plot to folder called "results"
ggsave(filename = "key_plot_7.png", 
      plot = plot7, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)


```

## Key plot 8

Kaplan Meier curve showing the probability of HC-free survival over time (n=562 months).

According to the plot below, around 200 months 50% of the patients were HC-free independently of the therapy received. All patients developed HC in around 550 months after the therapy ending.

```{r}
#Create a new data frame to add columns "event" and "time"
event_time <- 
  analysis_1.2 |> 
  mutate(
    event = as.numeric(hc_incidence == "no"), 
    time = rep(1:562, length.out = n())
  )

# Fit a Kaplan-Meier survival curve using the function: survfit
km_fit <- survfit(Surv(time, event) ~ 1, data = event_time)

# Create the survival plot
plot8 <- ggsurvplot(
  km_fit,
  xlab = "Time (Months)", 
  ylab = "HC-Free Survival Probability",
  legend.title = '',
  palette = "blue", 
  lwd = 2,
  conf.int = FALSE, 
  ylim = c(0, 1),
  legend.labs = "HC-Free Survival",
  legend = "right",
  theme = theme_minimal(base_size = 12)) +
  ggtitle("Kaplan-Meier curve of HC-Free Survival P. vs Time")

#Print plot
print(plot8)


# Save the plot to folder results
ggsave(
  filename = "key_plot_8.png", 
  plot = ggplot2::last_plot(),  
  path = "~/group_18_project/results",
  width = 8, 
  height = 5, 
  dpi = 300
)
```

## Key plot 9

Description: This plot illustrates the relationship between radiation dosage and duration across different types of cancer. Each cancer type is represented in a separate subplot for easy comparison, and the points are color-coded by radiation type. This visualization effectively displays complex data in a clear and comparative manner, ideal for in-depth analysis.

```{r warning = FALSE}

# Faceted Scatter Plot
plot9 <- ggplot(data = analysis_1.1, 
         aes(x = radiation_dosage, 
             y = radiation_duration, 
             color = radiation_type)) +
  geom_point() +
  facet_grid(. ~ cancer_type) +
  labs(title = "Faceted Scatter Plot of Radiation Dosage vs. Duration by Cancer Type",
       x = "Radiation Dosage",
       y = "Radiation Duration")

# Display the faceted scatter plot
print(plot9)

# Save the plot to folder results
ggsave(filename = "key_plot_9.png", 
      plot = plot9, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)


```

## Key plot 10

Description: The boxplot shows the amount of patients with HC and the different types of comorbidities that occured, while on the y-axis the amount of those comorbidities are shown. Only the cases with more than 1 comorbidity were taken into account

```{r}
# Separate hospitalization complications into distinct rows
df_comorbidities <- analysis_1.1 |>
  separate_rows(comorbidities, sep = ", ") |>
  filter(!is.na(comorbidities)) |>
  mutate(comorbidities = str_to_upper(comorbidities))  # Convert to uppercase

# Filter for complications that occurred in more than one person
df_filtered <- df_comorbidities |>
  group_by(comorbidities) |>
  filter(n() > 1) |>
  ungroup()

# Create a bar plot for the count of complications, colored by type of complication
plot10 <- ggplot(df_filtered, aes(x = comorbidities, fill = comorbidities)) +
  geom_bar(stat = "count", position = "dodge", color = "black", alpha = 0.5) +
  labs(title = "Count of Comorbidities (Cases with >1 comorbidity)",
       x = "Comorbidities",
       y = "Count",
       fill = "Comorbidities") +
  theme_minimal() +
  theme(axis.text.x = element_blank())  

#Print plot
print(plot10)

# Save the plot to folder results
ggsave(filename = "key_plot_10.png", 
      plot = plot10, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)
```

## Key plot 11

Description: The boxplot shows the amount of patients with HC and the hospital complications they experienced, while on the y-axis the number of hospital compilations are presented. Only the patients with more than 1 compilation are shown.

```{r}
# Create a new column with fixed hospitalization complications
df <- analysis_1.1 |>
  mutate(
    hospitalization_complications_fixed = case_when(
      row_number() %in% c(4, 28) ~ "Urosepsis, death",
      row_number() == 38 ~ "hydronephrosis requiring nephrostomy tubes, Urosepsis, right hydro requiring stent",
      row_number() == 31 ~ "ureteral strictures, hydronephrosis",
      row_number() == 65 ~ "AKI, anemia, prostatic necrosis, death",
      row_number() == 71 ~ "Bacteremia, PEG tube, death",
      row_number() == 15 ~ "tardive dyskinesia, AKI, elevation of PT",
      TRUE ~ as.character(hospitalization_complications)
    )
  )

# Separate hospitalization complications into distinct rows
df_complications <- df |>
  separate_rows(hospitalization_complications_fixed, sep = ", ") |>
  filter(!is.na(hospitalization_complications_fixed)) |>
  mutate(hospitalization_complications_fixed = str_to_upper(hospitalization_complications_fixed))  # Convert to uppercase

# Filter for complications that occurred in more than one person
df_filtered <- df_complications |>
  group_by(hospitalization_complications_fixed) |>
  filter(n() > 1) |>
  ungroup()

# Create a bar plot for the count of complications, colored by type of complication
plot11 <- ggplot(df_filtered, aes(x = hospitalization_complications_fixed, fill = hospitalization_complications_fixed)) +
  geom_bar(stat = "count", position = "dodge", color = "black", alpha = 0.5) +
  labs(title = "Count of Hospital Complications (Cases with >1 Complication)",
       x = "Hospital Complications",
       y = "Count",
       fill = "Hospitalization complications") +
  theme_minimal() +
  theme(axis.text.x = element_blank()) 

#Print plot
print(plot11)

# Save the plot to folder results
ggsave(filename = "key_plot_11.png", 
      plot = plot11, 
      path = "~/group_18_project/results",
      width = 8, 
      height = 5, 
      dpi = 300)

```
